requirements.txt
apache-airflow==2.8.3
requests
tenacity
pyspark==3.5.0
pytest
requests-mock
apache-airflow-providers-apache-spark==4.11.3
psycopg2-binary==2.9.9
